Refactored code in ad on backend.

+++ src/main/java/com/exposit/carsharing/service/AdServiceImpl.java
@@ -2,14 +2,12 @@ package com.exposit.carsharing.service;
 
 import com.exposit.carsharing.domain.Ad;
 import com.exposit.carsharing.domain.Car;
-import com.exposit.carsharing.domain.Profile;
 import com.exposit.carsharing.dto.AdRequest;
 import com.exposit.carsharing.dto.AdResponse;
 import com.exposit.carsharing.exception.EntityAlreadyExistException;
 import com.exposit.carsharing.exception.EntityNotFoundException;
 import com.exposit.carsharing.exception.PrivilegeException;
 import com.exposit.carsharing.repository.AdRepository;
-import com.exposit.carsharing.repository.CarRepository;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -24,19 +22,13 @@ public class AdServiceImpl implements AdService {
     private final AdRepository adRepository;
     private final CarService carService;
     private final ModelMapper modelMapper;
-    private final CarRepository carRepository;
 
-    public AdServiceImpl(ProfileService profileService, AdRepository adRepository, CarService carService, ModelMapper modelMapper, CarRepository carRepository) {
+    public AdServiceImpl(ProfileService profileService, AdRepository adRepository, CarService carService,
+                         ModelMapper modelMapper) {
         this.profileService = profileService;
         this.adRepository = adRepository;
         this.carService = carService;
         this.modelMapper = modelMapper;
-        this.carRepository = carRepository;
-    }
-
-    @Override
-    public boolean isExist(Long id) {
-        return adRepository.findOne(id) != null;
     }
 
     @Override
@@ -50,52 +42,80 @@ public class AdServiceImpl implements AdService {
 
     @Override
     public AdResponse getAdResponse(Long id) throws EntityNotFoundException {
-        return modelMapper.map(getAd(id), AdResponse.class);
+        return mapToResponse(getAd(id));
     }
 
     @Override
-    public List<Ad> getAll() {
-        return adRepository.findAll();
+    public List<AdResponse> getAll() {
+        return mapAllToResponse(adRepository.findAll());
     }
 
     @Override
-    public List<AdResponse> getAllAdByOwner(Long ownerId) throws EntityNotFoundException {
-        Profile owner = profileService.getProfile(ownerId);
-        List<AdResponse> ads = new ArrayList<>();
-        adRepository.findAllByOwner(owner).forEach(ad -> ads.add(modelMapper.map(ad, AdResponse.class)));
-        return ads;
+    public List<AdResponse> getAllByOwner(Long ownerId) throws EntityNotFoundException {
+        return mapAllToResponse(profileService.getProfile(ownerId).getAds());
     }
 
     @Override
-    public AdResponse createAd(AdRequest adRequest, Long ownerId, Long carId) throws EntityNotFoundException, EntityAlreadyExistException {
-        Ad ad = modelMapper.map(adRequest, Ad.class);
-        ad.setOwner(profileService.getProfile(ownerId));
+    public AdResponse create(AdRequest adRequest, Long ownerId, Long carId)
+            throws EntityNotFoundException, EntityAlreadyExistException, PrivilegeException {
         Car car = carService.getCar(carId);
+        checkCarOwner(ownerId, car);
+        checkAdExistForCar(car);
+        Ad ad = mapFromRequest(adRequest);
+        ad.setOwner(profileService.getProfile(ownerId));
         ad.setCar(carService.getCar(carId));
         adRepository.save(ad);
-        car.setAd(ad);
-        carRepository.save(car);
-        return modelMapper.map(ad, AdResponse.class);
+        return mapToResponse(ad);
     }
 
     @Override
-    public AdResponse updateAd(Long adId, AdRequest adRequest) throws EntityNotFoundException {
+    public AdResponse update(Long ownerId, Long adId, AdRequest adRequest)
+            throws EntityNotFoundException, PrivilegeException {
         Ad ad = getAd(adId);
+        checkAdOwner(ownerId, ad);
         ad.setCarLocation(adRequest.getCarLocation());
         ad.setReturnPlace(adRequest.getReturnPlace());
         ad.setCostPerHour(adRequest.getCostPerHour());
         ad.setCostPerDay(adRequest.getCostPerDay());
         ad.setCostPer3Days(adRequest.getCostPer3Days());
-        adRepository.save(ad);
-        return modelMapper.map(ad, AdResponse.class);
-
+        return mapToResponse(ad);
     }
 
     @Override
-    public void delete(Long adId) throws PrivilegeException, EntityNotFoundException {
-        /*if (!getAd(adId).getOwner().getId().equals(ownerId)) {
-            throw new PrivilegeException();
-        }*/
+    public void delete(Long ownerId, Long adId) throws PrivilegeException, EntityNotFoundException {
+        checkAdOwner(ownerId, getAd(adId));
         adRepository.delete(adId);
+    }
+
+    private void checkCarOwner(Long ownerId, Car car) throws PrivilegeException {
+        if (!car.getOwner().getId().equals(ownerId)) {
+            throw new PrivilegeException();
+        }
+    }
+
+    private void checkAdOwner(Long ownerId, Ad ad) throws PrivilegeException {
+        if (!ad.getOwner().getId().equals(ownerId)) {
+            throw new PrivilegeException();
+        }
+    }
+
+    private void checkAdExistForCar(Car car) throws EntityAlreadyExistException {
+        if (car.getAd() != null) {
+            throw new EntityAlreadyExistException(String.format("Ad for car with id %d already exist", car.getId()));
+        }
+    }
+
+    private Ad mapFromRequest(AdRequest adRequest) {
+        return modelMapper.map(adRequest, Ad.class);
+    }
+
+    private AdResponse mapToResponse(Ad ad) {
+        return modelMapper.map(ad, AdResponse.class);
+    }
+
+    private List<AdResponse> mapAllToResponse(List<Ad> ads) {
+        List<AdResponse> adResponses = new ArrayList<>();
+        ads.forEach(ad -> adResponses.add(mapToResponse(ad)));
+        return adResponses;
     }
 }
